Hybrid Core 1.3 already have a post format shortcode.
The only difference is that it links to the post format archive instead of just returning the post format.
So the entry meta for some of the post formats have been changed so there's always at least one permalink to the post.

The Cakifo shortcode [entry-format] is kept if some one just wants to display the post format name without a link