Split the subsidiary sidebar into three sidebars
This commit splits the subsidiary sidebar into three different. The name
has also been renamed to 'Footer Area One' and so on to make it easier
to understand

The theme auto detects if there's one, two or three active sidebars and
adjust the width after that. Just like in Twentyeleven