Use cakifo_logo() instead of hybrid_site_title()
The hybrid_site_title() function was "[..] created for XHTML themes and
a time when we needed to use an <h1> on the front page but an <h2> on
interior pages. With HTML5, we no longer need to do that." ~ Justin
Tadlock

Before cakifo_logo() was filtering the hybrid_site_title() function.
It's easier to just use that function as a replacement.