Set default background color
For some reason WordPress doesn't add the default background-color to
the head. So this sets the color in style.css and filters the body class
to add a custom-background-empty/white class to the color can be
removed.

custom-background-empty removes the box-shadow from the content wrapper
and footer.

Currently custom-background-white doesn't do anything