Better support for chat post format
This function filters the post content when viewing a post with the
"chat" post format.  It formats the content with structured HTML markup
to make it easy for theme developers to style chat posts.

The advantage of this solution is that it allows for more than two
speakers (like most solutions).

You can have 100s of speakers in your chat post, each with their own,
unique classes for styling.

See http://justintadlock.com/archives/2012/08/21/post-formats-chat