Improve navigation to be mobile-first
The navigation is now taking a mobile-first approach which means the
menu is hidden by default and have a big "Show menu" button

When the device width is above 980px the navigation is showed a usual.
Same way as before.

It simplifies the javascript needed and there's no need for the
Mediaqueries test in Modernizr anymore.