Add support for 3.6 structured post formats
'Link' format only for now. The other formats will be added when 3.6
beta 1 is ready as the code around post formats are changing fast these
hours. Hopefully it's very soon.

* Adds cakifo_get_link_url() to get the URL from the post. For now it
replaces post_format_tools_url_grabber() but we'll see what changes are
made in Hybrid Core in the coming time.