Remove border-radius from the content and footer.
Using both box-shadow and border-radius on an element can lead to very
expensive painting operations. In my testing, the paint time increased
from ~8 ms to ~30 ms.

The use of the border-radius was only a minor one as it only was applied
to the bottom left and right corner of the elements.

See
http://updates.html5rocks.com/2013/02/Profiling-Long-Paint-Times-with-DevTools-Continuous-Painting-Mode