Introduce cakifo_get_headline_term()
The function returns a Term object with info about a term selected in
the 'Headlines' option.

This also fixes a bug with the customizer as it stores The Customizer
the terms in a string in the `taxonomy:id` format.

The option is either an array, string in the `taxonomy:id` format, a
string with just the ID, or an int with the ID.