Turn the Headline lists into widgets
They were previously stored in a setting. But that caused many issues
with stuff like reordering and Chosen.js doesn’t play very well with
the Customizer.

This commit:
* Adds a widget-area to the front page
If the widget-area is inactive, the theme looks for the old setting and
shows the lists. That way the user doesn’t have to change anything when
the update the theme but if they want to update their columns, they now
have to use the widget.

* Adds the Headline Columns as a widget
Which means that they can have different settings such as a different
number of posts.

* Adds per-widget settings
- Thumbnail can be hidden
- Meta can be hidden
- Number of posts can be changed

* Removes the Headline Customizer settings and Chosen.js as they’re no
longer needed

* No longer removes duplicate posts from the list. A post can now be
both in the “Recent Posts” section and one or more Headline columns.